+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on Keep a Changelog, with one section per released version.
 
 ### Added
  - Business cards now show time *and* characters. Added toggle for choice of which metric to calculate percentages by.
+ - Added app startup workflow to recover a database with corrupted milestones
 
 ## [0.3.1] - 2026-07-24